LP-68: check err returned by http.Request.Cookie
The current auth template generates code that doesn't compile for
Cookie type openapi auth. I believe this dates from when this code
was rewritten from fastAPI (which has a Cookie function with no err
returned), but was never tested because cookie auth is not used.

This code gets and checks the returned error.