Copy headers one by one
This is a review comment.

Overwriting the headers throws away any that are set by middleware or
the http lib. It also depended on me mistakingly overwriting the
*request* headers instead of response headers.

This code sets headers one by one. It supports multiple of the same
header. It only sets headers that are defined to be set in the API spec.