Openapi: enforce 'required' from allOf/anyOf/oneOf
Consider the following OpenAPI spec:

    components:
      schema:
        ASchema:
          type: object
          properties:
            a:
              type: string
          required:
            - a
        BSchema:
          allOf:
            - $ref: '#/components/schemas/ASchema'
          type: object
          properties:
            b:
              type: string
          required:
            - b

An instance of BSchema must have both `a` and `b` fields set, however
currently foji only looks at BSchema.required and therefore only creates
validation for `b`.

This commit ensure we generate validation for fields marked required in
any `allOf` and for any fields that are marked as required in *all* of
`anyOf` and for any fields that are marked as required in *all* of
`oneOf`.