Choose Op Happy Response deterministically
If the openapi spec defines an authenticated operation with multiple 200-399 response codes (like we do in the "shortener" service) then foji picks a response non-deterministically. The root cause is that kin-openapi uses a map to store the responses and so we iterate them in random order.

This commit changes GetOpHappyResponse to sort the possible happy responses lexographically so that we always return the same one for any given input. It also means that when "200" is available, it is always chosen as the default even if other happy responses are available.