LP-520: Don't parse body when x-raw-request:true
x-raw-request currently passes the http.Request through to the handler
as-is. This works well when there is no body defined in the openapi
spec, which is how it has mostly been used to date. However when a body
is defined in the openapi spec the body is parsed as per the spec.
Beause r.Body() can only be read one time, this makes the raw request
useless for any purposes that would access this body.

This commit brings x-raw-request in line with the intended behaviour of
leaving request processing to the handler by omitting the body parsing.

This is a BREAKING CHANGE for any code that uses x-raw-request:true and
a defined body, as the signature and responsibility of the user code
changes.